Handle multiple values for an inferred event field: build multimaps when constructing inferred events. The state provider keeps a list of multi-valued inferred events.
Add a dialog to display the possible values for each multi-valued field of every inferred event.
The user can then choose a value by selecting it, and a specific structure is updated with the
selected value.

Add a button to the toolbar of the coherence view, which triggers the new dialog when clicked.