Create a new trace type, called InferenceTrace, which allows us to insert inferred events in a real CTF trace and run some analysis using these events.
Use this trace type in the GlobalInferenceView in order to display the
trace and its inferred events.

Infer some fields from the event name (and not only from the XML
conditions, as previously).
Select a value for a multi-valued field as the chosen one, if this
value comes from a particular inference where it was the only possible
one, while the other inferences computed several values. It indicates
this is the good value with high probability.